ops_rbac_user_spec - move user session methods to a helper
